Starting to break out the per-kind logic
